Improve system flexibility
Separate isLessThan and isGreaterThan functions into isLessThan, isLessThanOrEqualTo, isGreaterThan, and isGreaterThanOrEqualTo respectively.
For any functions that allow an AbstractLength or an AbstractArea as an argument, update the function to accept either 1 argument that is an input or 2 arguments that match the inputs to construct a new instance.
Update testing.
Bug fixes.